refactor(tests): split common/mod.rs and mocks.ts by domain, deduplicate settle_first_run_redirect
- Split `crates/e2e-tests/tests/common/mod.rs` (2907L) into four sub-modules:
  boot.rs (PrewarmedCache, InstanceEnv, port allocation, timeout consts),
  app.rs (E2eApp, InvokeOutcome, ResetScope, Drop impl),
  helpers.rs (process management, DB/storage reset, path resolution, ProcLog),
  fixtures.rs (FITS writers, scan_and_classify_one_item, settle_first_run_redirect).
  mod.rs becomes a re-export hub.

- Split `apps/desktop/src/api/mocks.ts` (3163L) domain data into
  src/api/mocks/{path-preview,audit,settings,targets,calibration,inbox,onboarding}.ts.
  mockHandlers registry, MockRegistry type, and mockInvoke remain in mocks.ts.
  Mock-baseline gate and registry anchors preserved.

- Deduplicate settle_first_run_redirect: 10 local copies removed from
  calibration_ui_journeys, inbox_ui_journeys, inventory_journeys,
  lifecycle_ui_journeys, onboarding_journey, plan_review_overlay_journeys,
  sessions_journeys, settings_journeys, source_view_journeys, targets_journeys;
  single canonical impl lives in common::fixtures.
