+++ crates/e2e-tests/tests/inbox_ui_journeys.rs
@@ -25,7 +25,7 @@ use std::path::Path;
 use std::time::Duration;
 
 use anyhow::Context;
-use common::{write_minimal_fits, write_minimal_fits_with_exposure, E2eApp};
+use common::{settle_first_run_redirect, write_minimal_fits, write_minimal_fits_with_exposure, E2eApp};
 use serde_json::json;
 
 const UI_TIMEOUT: Duration = Duration::from_secs(20);
@@ -70,22 +70,6 @@ async fn register_project_root(app: &E2eApp) -> anyhow::Result<tempfile::TempDir
         )
         .await?;
     Ok(project_dir)
-}
-
-/// Wait for the index route's async first-run redirect to land on `/setup`
-/// BEFORE navigating anywhere. A fresh DB (the harness resets it every
-/// launch) makes `checkFirstRunComplete` redirect `/` → `/setup` from an
-/// async `beforeLoad` (dynamic import + `firstrun_state` IPC round-trip,
-/// `apps/desktop/src/app/router.tsx`). If a journey `goto_route`s while that
-/// redirect is still pending, the late-resolving redirect yanks the app off
-/// the target route — on CI run 28766017315 that intermittently replaced
-/// `/#/inbox` with `/#/setup` and "Rescan all roots" never appeared. Once the
-/// URL shows `/setup`, no navigation is pending and `goto_route` is safe.
-async fn settle_first_run_redirect(app: &E2eApp) -> anyhow::Result<()> {
-    app.wait_url_contains("/setup", Duration::from_secs(15))
-        .await
-        .map(drop)
-        .map_err(|e| anyhow::anyhow!("expected a fresh DB to redirect to /setup: {e}"))
 }
 
 /// Seed the FIRST scan of `root_id` through the invoke bridge (a setup step,
