+++ crates/e2e-tests/tests/targets_journeys.rs
@@ -26,7 +26,7 @@ mod common;
 use std::time::Duration;
 
 use anyhow::Context;
-use common::E2eApp;
+use common::{settle_first_run_redirect, E2eApp};
 use serde_json::{json, Value};
 use thirtyfour::{By, WebElement};
 
@@ -45,20 +45,6 @@ const TARGETS_TABLE_TIMEOUT: Duration = Duration::from_secs(60);
 /// instead of silently drifting.
 const OPPOSITION_UNKNOWN_TITLE: &str = "Opposition date unknown — this target has no coordinates.";
 const LUNAR_UNKNOWN_TITLE: &str = "Lunar distance unknown — this target has no coordinates.";
-
-/// Wait for the index route's async first-run redirect to land on `/setup`
-/// BEFORE navigating anywhere (mirrors `inbox_ui_journeys.rs`'s
-/// `settle_first_run_redirect`). A fresh DB (the harness resets it every
-/// launch) makes `checkFirstRunComplete` redirect `/` → `/setup` from an
-/// async `beforeLoad`; if a journey `goto_route`s while that redirect is
-/// still pending, the late-resolving redirect can yank the app off the
-/// target route.
-async fn settle_first_run_redirect(app: &E2eApp) -> anyhow::Result<()> {
-    app.wait_url_contains("/setup", Duration::from_secs(15))
-        .await
-        .map(drop)
-        .map_err(|e| anyhow::anyhow!("expected a fresh DB to redirect to /setup: {e}"))
-}
 
 /// Complete first-run via the invoke bridge (native folder pickers can't be
 /// driven by WebDriver — same documented constraint as `journeys.rs`/
